removing the constant from the global collection
