+++ tina/collections/articles.tsx
@@ -1,3 +1,4 @@
+import { articlesIndexSchemaConstants } from "@/components/articles/articlesHeader";
 import type { Collection } from "tinacms";
 import * as Schemas from "../../components/blocks";
 import { carouselBlockSchema } from "../../components/blocks/carousel";
@@ -20,28 +21,6 @@ import {
 import { seoSchema } from "../../components/util/seo";
 import { sidebarPanelSchema } from "../../components/util/sidebarPanel";
 import { tipField } from "./shared-fields";
-
-export const articlesIndexSchemaConstants = {
-  value: "articlesIndex",
-  title: "title",
-  headerImage: {
-    value: "headerImage",
-    heroBackground: "heroBackground",
-    altText: "altText",
-    txtOverlay: "txtOverlay",
-  },
-  _body: "_body",
-  articles: {
-    value: "articles",
-    title: "title",
-    body: "body",
-    pageURL: "pageURL",
-    isExternal: "isExternal",
-    userName: "userName",
-    userPosition: "userPosition",
-    userImage: "userImage",
-  },
-};
 
 export const articlesSchema: Collection = {
   label: "Articles - Pages",
